+++ mne/beamformer/__init__.py
@@ -1,0 +1,4 @@
+"""Artifacts finding/correction related functions
+"""
+
+from ._lcmv import lcmv