+++ mne/fiff/tests/test_pick.py
@@ -8,3 +8,4 @@ def test_pick_channels_regexp():
     assert_array_equal(pick_channels_regexp(ch_names, 'MEG ...1'), [0])
     assert_array_equal(pick_channels_regexp(ch_names, 'MEG ...[2-3]'), [1, 2])
     assert_array_equal(pick_channels_regexp(ch_names, 'MEG *'), [0, 1, 2])
+