+++ doc/source/whats_new.rst
@@ -9,6 +9,10 @@ Changelog
 
    - Backporting scipy.signa.firwin2 so filtering works with old scipy by `Alex Gramfort`_.
 
+   - LCMV Beamformer by `Alex Gramfort`_.
+
+   - Add support for reading named channel selections by `Martin Luessi`_.
+
    - Add Raw.filter method to more easily band pass data by `Alex Gramfort`_.
 
    - Add tmin + tmax parameters in mne.compute_covariance to estimate noise covariance in epochs baseline without creating new epochs by `Alex Gramfort`_.